Use a slim production origin entrypoint
