+++ package.json
@@ -43,6 +43,7 @@
     "nx-cache:smoke": "node tools/nx-remote-cache-smoke.ts",
     "runtime:build": "node ./wrangler-env.ts deploy --dry-run --outdir .wrangler/runtime-dry-run --config packages/runtime-worker/wrangler.jsonc",
     "platform:build": "node ./wrangler-env.ts deploy --dry-run --outdir .wrangler/platform-dry-run --config packages/platform-worker/wrangler.jsonc",
+    "main:build": "node ./wrangler-env.ts deploy --dry-run --outdir .wrangler/main-dry-run --env production",
     "jobs:build": "wrangler deploy --dry-run --outdir .tmp/jobs-worker-build --config packages/jobs-worker/wrangler.jsonc --env production",
     "jobs:deploy": "wrangler deploy --config packages/jobs-worker/wrangler.jsonc --env production",
     "highlight:build": "wrangler deploy --dry-run --outdir .tmp/highlight-worker-build --config packages/highlight-worker/wrangler.jsonc --env production",
@@ -59,6 +60,7 @@
     "docs:check-temporal": "node tools/check-docs-temporal-language.ts",
     "docs:check-decisions": "node tools/check-decision-record-numbers.ts",
     "deploy-guardrails:check": "node tools/check-deploy-guardrails.ts",
+    "origin-production-exports:check": "node tools/check-origin-production-exports.ts",
     "migrations:check": "node tools/check-migrations.ts",
     "migrate:local": "node tools/apply-local-app-migrations.ts && node --env-file=packages/worker/.env ./wrangler-env.ts d1 migrations apply AUDIT_DB --local && node --env-file=packages/worker/.env ./wrangler-env.ts d1 migrations apply JOBS_DB --local --config packages/jobs-worker/wrangler.jsonc",
     "migrate:e2e": "node tools/apply-local-app-migrations.ts --persist-to .wrangler/state/e2e && node --env-file=packages/worker/.env ./wrangler-env.ts d1 migrations apply AUDIT_DB --local --persist-to .wrangler/state/e2e && node --env-file=packages/worker/.env ./wrangler-env.ts d1 migrations apply JOBS_DB --local --config packages/jobs-worker/wrangler.jsonc --persist-to .wrangler/state/e2e",
@@ -74,7 +76,7 @@
     "test:workers": "nx run worker:test-workers",
     "test:push": "CI=1 npm run test:node && CI=1 npm run test:workers",
     "inspect": "npx -y @mcpjam/inspector inspector",
-    "validate": "concurrently -n format,lint,typecheck,test-node,test-workers,e2e,mcp,backup-build,status-build,nx-cache-build,jobs-build,highlight-build,runtime-build,platform-build,primitives,migrations,deploy-guardrails,docs-temporal,docs-decisions -c green,yellow,magenta,blue,cyan,red,blueBright,yellowBright,cyanBright,cyan,white,gray,redBright,greenBright,magentaBright,whiteBright,black,white,white \"npm run format:check\" \"npm run lint\" \"npm run typecheck\" \"CI=1 npm run test:node\" \"CI=1 npm run test:workers\" \"CI=1 npm run test:e2e:run\" \"CI=1 npm run test:mcp\" \"npm run backup:build\" \"npm run status:build\" \"npm run nx-cache:build\" \"npm run jobs:build\" \"npm run highlight:build\" \"npm run runtime:build\" \"npm run platform:build\" \"npm run primitives:check\" \"npm run migrations:check\" \"npm run deploy-guardrails:check\" \"npm run docs:check-temporal\" \"npm run docs:check-decisions\"",
+    "validate": "concurrently -n format,lint,typecheck,test-node,test-workers,e2e,mcp,backup-build,status-build,nx-cache-build,jobs-build,highlight-build,runtime-build,platform-build,main-build,primitives,migrations,deploy-guardrails,origin-production-exports,docs-temporal,docs-decisions -c green,yellow,magenta,blue,cyan,red,blueBright,yellowBright,cyanBright,cyan,white,gray,redBright,greenBright,magentaBright,whiteBright,black,white,gray,white,cyan \"npm run format:check\" \"npm run lint\" \"npm run typecheck\" \"CI=1 npm run test:node\" \"CI=1 npm run test:workers\" \"CI=1 npm run test:e2e:run\" \"CI=1 npm run test:mcp\" \"npm run backup:build\" \"npm run status:build\" \"npm run nx-cache:build\" \"npm run jobs:build\" \"npm run highlight:build\" \"npm run runtime:build\" \"npm run platform:build\" \"npm run main:build\" \"npm run primitives:check\" \"npm run migrations:check\" \"npm run deploy-guardrails:check\" \"npm run origin-production-exports:check\" \"npm run docs:check-temporal\" \"npm run docs:check-decisions\"",
     "validate:fix": "npm run format && npm run lint:fix",
     "test:e2e:ensure": "node tools/ensure-playwright-browser.ts",
     "test:e2e:a11y": "playwright test e2e/a11y.spec.ts",
